+++ web/src/pages/settings/acl/SettingsACL.vue
@@ -6,27 +6,32 @@
 
     <template #default>
       <div class="max-w-7xl">
-        <Header>Team and collaborators</Header>
-        <SettingsCollaborators :codebase-id="data.codebase.id" />
+        <Header>Security and ACLs</Header>
+        <SettingsACL
+          v-if="data?.codebase?.acl?.id"
+          :codebase-id="data.codebase.id"
+          :acl-id="data.codebase.acl.id"
+          :acl-policy="data.codebase.acl.policy"
+        />
       </div>
     </template>
   </PaddedAppLeftSidebar>
 </template>
 
 <script>
-import SettingsCollaborators from '../components/codebase/settings/SettingsCollaborators.vue'
 import { gql, useQuery } from '@urql/vue'
 import { useRoute } from 'vue-router'
-import PaddedAppLeftSidebar from '../layouts/PaddedAppLeftSidebar.vue'
-import SettingsVerticalNavigation from '../components/codebase/settings/SettingsVerticalNavigation.vue'
-import Header from '../molecules/Header.vue'
+import PaddedAppLeftSidebar from '../../../layouts/PaddedAppLeftSidebar.vue'
+import SettingsVerticalNavigation from '../../../components/codebase/settings/SettingsVerticalNavigation.vue'
+import Header from '../../../molecules/Header.vue'
+import SettingsACL from '../../../components/codebase/settings/SettingsACL.vue'
 
 export default {
   name: 'CodebaseSettings',
   components: {
+    SettingsACL,
     PaddedAppLeftSidebar,
     SettingsVerticalNavigation,
-    SettingsCollaborators,
     Header,
   },
   setup() {
@@ -34,9 +39,13 @@ export default {
 
     let { data } = useQuery({
       query: gql`
-        query SettingsTeam($id: ID, $shortID: ID) {
+        query SettingsAclPage($id: ID, $shortID: ID) {
           codebase(id: $id, shortID: $shortID) {
             id
+            acl {
+              id
+              policy
+            }
           }
         }
       `,
