+++ web/src/pages/settings/workspaces/SettingsWorkspaces.vue
@@ -5,14 +5,9 @@
     </template>
 
     <template #default>
-      <div class="max-w-7xl">
-        <Header>Security and ACLs</Header>
-        <SettingsACL
-          v-if="data?.codebase?.acl?.id"
-          :codebase-id="data.codebase.id"
-          :acl-id="data.codebase.acl.id"
-          :acl-policy="data.codebase.acl.policy"
-        />
+      <div class="max-w-7xl space-y-4">
+        <Header>Restore Workspaces</Header>
+        <SettingsWorkspaces classs="mt-8" :codebase-id="data.codebase.id" />
       </div>
     </template>
   </PaddedAppLeftSidebar>
@@ -21,15 +16,15 @@
 <script>
 import { gql, useQuery } from '@urql/vue'
 import { useRoute } from 'vue-router'
-import PaddedAppLeftSidebar from '../layouts/PaddedAppLeftSidebar.vue'
-import SettingsVerticalNavigation from '../components/codebase/settings/SettingsVerticalNavigation.vue'
-import Header from '../molecules/Header.vue'
-import SettingsACL from '../components/codebase/settings/SettingsACL.vue'
+import PaddedAppLeftSidebar from '../../../layouts/PaddedAppLeftSidebar.vue'
+import SettingsVerticalNavigation from '../../../components/codebase/settings/SettingsVerticalNavigation.vue'
+import Header from '../../../molecules/Header.vue'
+import SettingsWorkspaces from '../../../components/codebase/settings/SettingsWorkspaces.vue'
 
 export default {
   name: 'CodebaseSettings',
   components: {
-    SettingsACL,
+    SettingsWorkspaces,
     PaddedAppLeftSidebar,
     SettingsVerticalNavigation,
     Header,
@@ -39,13 +34,9 @@ export default {
 
     let { data } = useQuery({
       query: gql`
-        query SettingsAclPage($id: ID, $shortID: ID) {
+        query SettingsWorkspaces($id: ID, $shortID: ID) {
           codebase(id: $id, shortID: $shortID) {
             id
-            acl {
-              id
-              policy
-            }
           }
         }
       `,
